[FIX] #77 Util 함수 분리
