+++ app/src/main/java/com/timi/seulseul/presentation/splash/SplashActivity.kt
@@ -1,26 +1,15 @@
 package com.timi.seulseul.presentation.splash
 
-import android.Manifest
-import android.app.Dialog
-import android.content.Context
 import android.content.Intent
-import android.content.pm.PackageManager
-import android.net.ConnectivityManager
-import android.net.NetworkCapabilities
-import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.WindowManager
-import android.widget.Button
 import android.widget.TextView
 import androidx.appcompat.app.AlertDialog
-import androidx.core.app.NotificationManagerCompat
 import androidx.core.content.ContextCompat
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import androidx.databinding.DataBindingUtil
 import com.timi.seulseul.MainApplication.Companion.prefs
 import com.timi.seulseul.R
-import com.timi.seulseul.databinding.ActivityOnBoardingBinding
+import com.timi.seulseul.Utils
 import com.timi.seulseul.presentation.MainActivity
 import com.timi.seulseul.presentation.onboarding.OnBoardingActivity
 import com.timi.seulseul.presentation.permission.PermissionActivity
@@ -45,7 +34,7 @@ class SplashActivity : AppCompatActivity() {
 
     private fun checkNetworkConnection() {
         // 네트워크가 사용불가능일 경우
-        if (!isNetworkAvailable(this)) {
+        if (!Utils.isNetworkAvailable(this)) {
             // AlertDialog 객체 생성
             val dialog = AlertDialog.Builder(this)
                 .setView(R.layout.dialog_network_check)
@@ -72,7 +61,7 @@ class SplashActivity : AppCompatActivity() {
 
             if (checkOnBoarding()) {
                 // 권한이 전부 다 허용됐을 때
-                if (checkPermission()) {
+                if (Utils.checkPermission(this)) {
                     startActivity(Intent(this, MainActivity::class.java))
                 } else {
                     // 권한 허용 안된게 존재할 때
@@ -88,31 +77,6 @@ class SplashActivity : AppCompatActivity() {
         }
     }
 
-    private fun isNetworkAvailable(context: Context): Boolean {
-        // [API 23이상] ConnectivityManager, NetworkCapabilities
-        val connectivityManager =
-            context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager?
-        if (connectivityManager != null) {
-
-            val capabilities =
-                connectivityManager.getNetworkCapabilities(connectivityManager.activeNetwork)
-            if (capabilities != null) {
-                // 셀룰러 체크
-                if (capabilities.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR)) {
-                    return true
-                }
-                // WIFI 체크
-                else if (capabilities.hasTransport(NetworkCapabilities.TRANSPORT_WIFI)) {
-                    return true
-                }
-                // 이더넷 체크
-                else if (capabilities.hasTransport(NetworkCapabilities.TRANSPORT_ETHERNET)) {
-                    return true
-                }
-            }
-        }
-        return false
-    }
 
     private fun checkOnBoarding(): Boolean {
         // sp에서 KEY_ONBOARDING 값체크, 값이 없다면 2번째 인자인 false가 기본값
@@ -127,41 +91,5 @@ class SplashActivity : AppCompatActivity() {
         return false
     }
 
-    private fun checkPermission(): Boolean {
-        // FINE_LOCATION 체크
-        val hasFineLocationPermission = ContextCompat.checkSelfPermission(
-            this, Manifest.permission.ACCESS_FINE_LOCATION
-        ) == PackageManager.PERMISSION_GRANTED
-        // COARSE_LOCATION 체크
-        val hasCoarseLocationPermission = ContextCompat.checkSelfPermission(
-            this, Manifest.permission.ACCESS_COARSE_LOCATION
-        ) == PackageManager.PERMISSION_GRANTED
-        // BACKGROUND_LOCATION 체크
-        val hasBackgroundLocationPermission =
-            // API 26(Q) 이상일 때
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-                ContextCompat.checkSelfPermission(
-                    this, Manifest.permission.ACCESS_BACKGROUND_LOCATION
-                ) == PackageManager.PERMISSION_GRANTED
-            } else {
-                // API 26미만이면 백그라운드가 포그라운드 권한에 종속됨
-                true
-            }
-
-
-        val hasNotificationPermission = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            // POST_NOTIFICATIONS 체크, API33 알림 퍼미션
-            ContextCompat.checkSelfPermission(
-                this, Manifest.permission.POST_NOTIFICATIONS
-            ) == PackageManager.PERMISSION_GRANTED
-        } else {
-            // Notification 허용 여부 체크 (모든 알림(채널) 차단만 체크해줌)
-            NotificationManagerCompat.from(this).areNotificationsEnabled()
-            // 특정 채널 알림 차단여부는 체크따로 해야함
-        }
-
-        // 전부다 true일 때 checkPermission을 true로 반환, 하나라도 false가 있으면 false로 반환
-        return hasFineLocationPermission && hasCoarseLocationPermission && hasBackgroundLocationPermission && hasNotificationPermission
-    }
 
 }
