+++ app/src/main/res/layout/activity_on_boarding.xml
@@ -17,7 +17,8 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.onboarding.OnBoardingActivity">
+        tools:context=".presentation.onboarding.OnBoardingActivity"
+        android:background="@color/white">
 
 
         <androidx.viewpager2.widget.ViewPager2
