+++ app/src/main/res/layout/dialog_network_check.xml
@@ -7,7 +7,7 @@
     android:layout_height="wrap_content"
     android:background="@drawable/bg_white_r8"
     android:gravity="center"
-    android:minWidth="300dp"
+    android:minWidth="340dp"
     android:minHeight="100dp"
     android:orientation="vertical"
     android:paddingHorizontal="@dimen/padding_20"
