+++ app/src/main/res/layout/item_on_boarding.xml
@@ -22,7 +22,7 @@
             android:layout_height="0dp"
             app:layout_constraintHeight_percent="0.7"
             app:imageResource="@{page.imageResId}"
-            android:scaleType="fitXY"
+            android:scaleType="centerCrop"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent"
